implement encoder-decoder (T5) support
